Hopefully fix `basePath` behaviour
